More test refinements.
Getting Python3 to work with text correctly is tricky.
Files contain bytes. Programs read and write bytes.
But Python insists on converting things to and from unicode strings.
Trying to leave everything as bytes as much as possible,
but still need to convert to Unicode to write results to sys.stdout.